[FEAT] 날씨, 인구 데이터 한번에 가져오기

+++ src/main/java/com/example/whatseoul/entity/Population.java
@@ -17,7 +17,7 @@ public class Population {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "POPULATION_ID")
-    private int populationId;
+    private Long populationId;
 
     @OneToOne
     @JoinColumn(name = "AREA_CODE")
